Move numeric testing to conversion and module level skipping in conversion testing
